Create Plot_all_evt.py
The code plots all events instead of the reconstructed events. 
It takes a long time to load all of the data so, it prints a line for every data file loaded to prevent disconnection. 
The data it reads is at home/genec420/gain3, however, some plots are behaving weirdly (showing two linear trend) and I'm suspecting something is wrong with the creating of the data.